Commit Johnson's all pair shortest path algorithm

+++ src/main/java/com/techstuff/algods/graphs/ComparableVertex.java
@@ -2,22 +2,22 @@ package com.techstuff.algods.graphs;
 
 public class ComparableVertex<T> extends Vertex<T> implements Comparable<ComparableVertex<T>> {
 	
-	private Integer key;
+	private Double key;
 
 	public ComparableVertex(T payload) {
 		super(payload);
 	}
 	
-	public Integer getKey() {
+	public Double getKey() {
 		return key;
 	}
 
-	public void setKey(Integer key) {
+	public void setKey(Double key) {
 		this.key = key;
 	}
 
 	@Override
 	public int compareTo(ComparableVertex<T> o) {
-		return this.key - o.key;
+		return (int)(this.key - o.key);
 	}
 }
